Move `section_name` builder to `common`
The `common/section_name.h` was not added to `common.h` because it's not
expected to be a default choice for every single test.
